UPDATE: Fixing some more namespace issues.
git-svn-id: svn://svn.code.sf.net/p/nel/code/branches/branch_mtr_nostlport@270 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/include/nel/misc/mem_stream.h
@@ -147,7 +147,7 @@ public:
 	CMemStream( bool inputStream=false, bool stringmode=false, uint32 defaultcapacity=0 ) :
 		NLMISC::IStream( inputStream ), _StringMode( stringmode )
 	{
-		_DefaultCapacity = std::max( defaultcapacity, (uint32)16 ); // prevent from no allocation
+		_DefaultCapacity = std::max( (uint32)defaultcapacity, (uint32)16 ); // prevent from no allocation
 		_Buffer.getBufferWrite().resize (_DefaultCapacity);
 		_Buffer.Pos = 0;
 	}
